pyproject.toml
[project]
name = "blog_api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=25.1.0",
    "bleach>=6.3.0",
    "django>=6.0.1",
    "django-autoslug>=1.9.9",
    "django-ratelimit>=4.1.0",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt[crypto]>=5.5.1",
    "loguru>=0.7.3",
    "pillow>=12.1.0",
    "psycopg[binary]>=3.3.2",
    "python-decouple>=3.8",
    "redis>=7.1.0",
    "zxcvbn>=4.5.0",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "django-stubs[compatible-mypy]>=5.2.9",
    "faker>=40.4.0",
    "mypy>=1.19.1",
    "ruff>=0.14.14",
]
[tool.django-stubs]
django_settings_module = "settings"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = []

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
ignore_missing_imports = true

python_version = "3.12"
strict = false
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
